cleanup, fix some object list bugs
